test(ui): cover safeExternalUrl SSRF barrier in lib
Closes #3431 — move the security barrier to lib/metagraphed and add
focused Vitest coverage for public http(s), credentialed, and private hosts.
